requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
deep-translator==1.11.4
Django==5.0.5
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
factory-boy==3.3.0
Faker==25.2.0
filelock==3.14.0
fsspec==2024.5.0
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-translate==3.15.3
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
gunicorn==22.0.0
huggingface-hub==0.23.2
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
Markdown==3.6
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
mysqlclient==2.2.4
networkx==3.3
numpy==1.26.4
packaging==24.0
pandas==2.2.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.2
rsa==4.9
safetensors==0.4.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
sympy==1.12
tbb==2021.12.0
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.41.1
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
